Fix error with focusing on form after dialog closes
The default nature of the form button to add entries
is to focus back on the first form element once it
clears all the previous book entry's data and closes
the dialog box. Adding the event.preventDefault()
method prevents this and fixes the error that occurs
when the browser window attempts to focus on the form
element after the dialog box has already closed.