Split functionality that displays logs into separate function
Having just one function that creates and displays
the library table logs (addBookToLibrary) makes the
function itself bulky and hard to understand. Splitting
the function code into two distinct functions for
adding and display the library table logs makes for
clearer code and greater flexibility in accessing the
book objects and their properties, as we are not
saving the entries in a list as table row elements
anymore.